// src/anytype-openapi/types.gen.ts
export type CheckboxPropertyLinkValue = {
    /**
     * The checkbox value of the property
     */
    checkbox?: boolean;
    key?: string;
};

export type CheckboxPropertyValue = {
    /**
     * The checkbox value of the property
     */
    checkbox?: boolean;
    format?: PropertyFormat;
    /**
     * The id of the property
     */
    id?: string;
    /**
     * The key of the property
     */
    key?: string;
    /**
     * The name of the property
     */
    name?: string;
    /**
     * The data model of the object
     */
    object?: string;
};

/**
 * The color of the icon
 */
export type Color = 'grey' | 'yellow' | 'orange' | 'red' | 'pink' | 'purple' | 'blue' | 'ice' | 'teal' | 'lime';
// ...
export type DatePropertyLinkValue = {
    /**
     * The date value of the property. Accepts dates in RFC3339 format (2006-01-02T15:04:05Z) or date-only format (2006-01-02)
     */
    date?: string;
    key?: string;
};

export type DatePropertyValue = {
    /**
     * The date value of the property. Returns dates in RFC3339 format (2006-01-02T15:04:05Z)
     */
    date?: string;
    format?: PropertyFormat;
    /**
     * The id of the property
     */
    id?: string;
    /**
     * The key of the property
     */
    key?: string;
    /**
     * The name of the property
     */
    name?: string;
    /**
     * The data model of the object
     */
    object?: string;
};

export type EmailFilterItem = {
    condition?: FilterCondition;
    /**
     * The email value to filter by
     */
    email?: string;
    /**
     * The property key to filter on
     */
    property_key?: string;
};

export type EmailPropertyLinkValue = {
    /**
     * The email value of the property
     */
    email?: string;
    key?: string;
};

export type EmailPropertyValue = {
    /**
     * The email value of the property
     */
    email?: string;
    format?: PropertyFormat;
    /**
     * The id of the property
     */
    id?: string;
    /**
     * The key of the property
     */
    key?: string;
    /**
     * The name of the property
     */
    name?: string;
    /**
     * The data model of the object
     */
    object?: string;
};
// ...
export type FilesPropertyLinkValue = {
    /**
     * The file ids of the property
     */
    files?: Array<string>;
    key?: string;
};

export type FilesPropertyValue = {
    /**
     * The file values of the property
     */
    files?: Array<string>;
    format?: PropertyFormat;
    /**
     * The id of the property
     */
    id?: string;
    /**
     * The key of the property
     */
    key?: string;
    /**
     * The name of the property
     */
    name?: string;
    /**
     * The data model of the object
     */
    object?: string;
};
// ...
export type MultiSelectPropertyLinkValue = {
    key?: string;
    /**
     * The selected tags (by key, e.g., "important", or ID, e.g., "bafyrei...") of the property; see ListTags endpoint for valid values
     */
    multi_select?: Array<string>;
};

export type MultiSelectPropertyValue = {
    format?: PropertyFormat;
    /**
     * The id of the property
     */
    id?: string;
    /**
     * The key of the property
     */
    key?: string;
    /**
     * The selected tag values of the property
     */
    multi_select?: Array<Tag>;
    /**
     * The name of the property
     */
    name?: string;
    /**
     * The data model of the object
     */
    object?: string;
};

export type NamedIcon = {
    color?: Color;
    format?: IconFormat;
    name?: IconName;
};

export type NumberFilterItem = {
    condition?: FilterCondition;
    /**
     * The number value to filter by
     */
    number?: number;
    /**
     * The property key to filter on
     */
    property_key?: string;
};

export type NumberPropertyLinkValue = {
    key?: string;
    /**
     * The number value of the property
     */
    number?: number;
};

export type NumberPropertyValue = {
    format?: PropertyFormat;
    /**
     * The id of the property
     */
    id?: string;
    /**
     * The key of the property
     */
    key?: string;
    /**
     * The name of the property
     */
    name?: string;
    /**
     * The number value of the property
     */
    number?: number;
    /**
     * The data model of the object
     */
    object?: string;
};
// ...
export type ObjectsPropertyLinkValue = {
    key?: string;
    /**
     * The object ids of the property
     */
    objects?: Array<string>;
};

export type ObjectsPropertyValue = {
    format?: PropertyFormat;
    /**
     * The id of the property
     */
    id?: string;
    /**
     * The key of the property
     */
    key?: string;
    /**
     * The name of the property
     */
    name?: string;
    /**
     * The data model of the object
     */
    object?: string;
    /**
     * The object values of the property
     */
    objects?: Array<string>;
};

export type PhoneFilterItem = {
    condition?: FilterCondition;
    /**
     * The phone value to filter by
     */
    phone?: string;
    /**
     * The property key to filter on
     */
    property_key?: string;
};

export type PhonePropertyLinkValue = {
    key?: string;
    /**
     * The phone value of the property
     */
    phone?: string;
};

export type PhonePropertyValue = {
    format?: PropertyFormat;
    /**
     * The id of the property
     */
    id?: string;
    /**
     * The key of the property
     */
    key?: string;
    /**
     * The name of the property
     */
    name?: string;
    /**
     * The data model of the object
     */
    object?: string;
    /**
     * The phone value of the property
     */
    phone?: string;
};
// ...
export type PropertyFormat = 'text' | 'number' | 'select' | 'multi_select' | 'date' | 'files' | 'checkbox' | 'url' | 'email' | 'phone' | 'objects';

export type PropertyLink = {
    format: PropertyFormat;
    /**
     * The key of the property
     */
    key: string;
    /**
     * The name of the property
     */
    name: string;
};

export type PropertyLinkWithValue = TextPropertyLinkValue | NumberPropertyLinkValue | SelectPropertyLinkValue | MultiSelectPropertyLinkValue | DatePropertyLinkValue | FilesPropertyLinkValue | CheckboxPropertyLinkValue | UrlPropertyLinkValue | EmailPropertyLinkValue | PhonePropertyLinkValue | ObjectsPropertyLinkValue;

export type PropertyResponse = {
    property?: Property;
};

export type PropertyWithValue = TextPropertyValue | NumberPropertyValue | SelectPropertyValue | MultiSelectPropertyValue | DatePropertyValue | FilesPropertyValue | CheckboxPropertyValue | UrlPropertyValue | EmailPropertyValue | PhonePropertyValue | ObjectsPropertyValue;
// ...
export type SelectPropertyLinkValue = {
    key?: string;
    /**
     * The selected tag (by key, e.g., "important", or ID, e.g., "bafyrei...") of the property; see ListTags endpoint for valid values
     */
    select?: string;
};

export type SelectPropertyValue = {
    format?: PropertyFormat;
    /**
     * The id of the property
     */
    id?: string;
    /**
     * The key of the property
     */
    key?: string;
    /**
     * The name of the property
     */
    name?: string;
    /**
     * The data model of the object
     */
    object?: string;
    select?: Tag;
};
// ...
export type Tag = {
    color?: Color;
    /**
     * The id of the tag
     */
    id?: string;
    /**
     * The key of the tag
     */
    key?: string;
    /**
     * The name of the tag
     */
    name?: string;
    /**
     * The data model of the object
     */
    object?: string;
};

export type TagResponse = {
    tag?: Tag;
};

export type TemplateResponse = {
    template?: ObjectWithBody;
};

export type TextFilterItem = {
    condition?: FilterCondition;
    /**
     * The property key to filter on
     */
    property_key?: string;
    /**
     * The text value to filter by
     */
    text?: string;
};

export type TextPropertyLinkValue = {
    key?: string;
    /**
     * The text value of the property
     */
    text?: string;
};

export type TextPropertyValue = {
    format?: PropertyFormat;
    /**
     * The id of the property
     */
    id?: string;
    /**
     * The key of the property
     */
    key?: string;
    /**
     * The name of the property
     */
    name?: string;
    /**
     * The data model of the object
     */
    object?: string;
    /**
     * The text value of the property
     */
    text?: string;
};
// ...
export type UrlPropertyLinkValue = {
    key?: string;
    /**
     * The URL value of the property
     */
    url?: string;
};

export type UrlPropertyValue = {
    format?: PropertyFormat;
    /**
     * The id of the property
     */
    id?: string;
    /**
     * The key of the property
     */
    key?: string;
    /**
     * The name of the property
     */
    name?: string;
    /**
     * The data model of the object
     */
    object?: string;
    /**
     * The URL value of the property
     */
    url?: string;
};